ci: refactor add item to project script to use env vars
